SSOR-276 Moved account service to boss-core

+++ src/main/java/org/ssor/boss/core/transfer/TransactionInput.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class TransactionInput
 {
   @NonNull
-  private Integer accountId;
+  private Long accountId;
 
   @Min(0)
   private Float amount;
